claude authored, updated boilerplate template for (minimal) machines
