+++ machine_implementations/src/minimal_machines/BOILERPLATE_TEMPLATE/api.rs
@@ -1,12 +1,12 @@
 // ============================================================================
-// api.rs — Events, mutations, and MachineApi implementation
+// api.rs — Events, Mutations, Namespace, MachineApi impl
 // ============================================================================
-// This file defines:
-//   • StateEvent      — the data structure broadcast to frontend subscribers
-//   • MyMachineEvents — enum wrapping all event types for this machine
-//   • Mutation        — the actions the frontend can send to this machine
-//   • MyMachineNamespace — the socket.io namespace handle
-//   • MachineApi impl — wires mutations and namespace access
+// Contains:
+//   • StateEvent       — payload broadcast to socket.io subscribers
+//   • MyMachineEvents  — enum wrapping every event variant this machine emits
+//   • Mutation         — actions the frontend can POST to this machine
+//   • MyMachineNamespace — socket.io namespace handle (mandatory plumbing)
+//   • MachineApi impl  — message dispatch + mutation handling
 // ============================================================================
 
 use std::sync::Arc;
@@ -20,61 +20,55 @@ use control_core::socketio::{
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use crate::{MachineApi, MachineMessage};
 use super::MyMachine;
+use crate::{MachineApi, MachineMessage, MachineValues};
 
 // ----------------------------------------------------------------------------
-// Step 1 — Define the state that gets broadcast to the frontend.
-//
-// Add every field that the UI needs to display. Must be Serialize + Clone.
+// Step 1 — Define the state broadcast to the frontend.
+// Every field the UI displays goes here. Must be Serialize + Clone.
 // ----------------------------------------------------------------------------
 #[derive(Serialize, Debug, Clone)]
 pub struct StateEvent {
-    // TODO: add your state fields, e.g.:
-    //   pub outputs: [bool; 4],
-    //   pub inputs:  [bool; 4],
-    //   pub value:   f64,
+    // TODO: state fields, e.g.:
+    // pub outputs: [bool; 4],
+    // pub inputs:  [bool; 4],
+    // pub value:   f64,
 }
 
 impl StateEvent {
-    /// Wrap state in a named socket.io event ready to be emitted.
     pub fn build(&self) -> Event<Self> {
         Event::new("StateEvent", self.clone())
     }
 }
 
 // ----------------------------------------------------------------------------
-// Step 2 — Declare all event variants your machine can emit.
-//
-// For most minimal machines a single `State` variant is enough.
-// Add more variants (e.g. `LiveValues`, `Alert`) only if the frontend needs
-// them distinguished.
+// Step 2 — Event variants this machine can emit.
+// One `State` variant is enough for most minimal machines. Add more (e.g.
+// `LiveValues`, `Alert`) only if the frontend needs to distinguish them.
 // ----------------------------------------------------------------------------
 pub enum MyMachineEvents {
     State(Event<StateEvent>),
-    // TODO: add more variants if needed, e.g.:
-    //   LiveValues(Event<LiveValuesEvent>),
+    // TODO: add more variants if needed.
 }
 
 // ----------------------------------------------------------------------------
-// Step 3 — Define mutations the frontend can trigger.
+// Step 3 — Frontend → backend mutations.
 //
-// Use `#[serde(tag = "action", content = "value")]` so the wire format is:
-//   { "action": "SetOutput", "value": { "index": 0, "on": true } }
+// Wire format with `#[serde(tag = "action", content = "value")]`:
+//   { "action": "SetOutput", "value": { "index": 0, "value": true } }
 //
-// If your machine is read-only (inputs only), keep the enum empty or with a
-// single never-used variant, and return Ok(()) in api_mutate.
+// If the machine is read-only, leave the enum empty AND remove the
+// `serde_json::from_value` call in `api_mutate` below (returning Ok(()) directly).
 // ----------------------------------------------------------------------------
 #[derive(Deserialize)]
 #[serde(tag = "action", content = "value")]
 pub enum Mutation {
-    // TODO: add your mutations, e.g.:
-    //   SetOutput { index: usize, on: bool },
-    //   SetAllOutputs { on: bool },
+    // TODO: mutations, e.g.:
+    // SetOutput { index: usize, value: bool },
 }
 
 // ----------------------------------------------------------------------------
-// Step 4 — The namespace handle (mandatory plumbing, do not change the struct).
+// Step 4 — Namespace handle (mandatory plumbing).
 // ----------------------------------------------------------------------------
 #[derive(Debug, Clone)]
 pub struct MyMachineNamespace {
@@ -95,36 +89,62 @@ impl CacheableEvents<MyMachineEvents> for MyMachineEvents {
     fn event_value(&self) -> GenericEvent {
         match self {
             MyMachineEvents::State(event) => event.clone().into(),
-            // TODO: add arms for every new variant you add above
+            // TODO: add an arm for every new variant.
         }
     }
 
     fn event_cache_fn(&self) -> CacheFn {
-        // `cache_first_and_last_event` keeps the very first and the most recent
-        // event in the cache so new subscribers always get a value immediately.
+        // Caches the first AND most recent event so new subscribers get a
+        // value immediately on connect.
         cache_first_and_last_event()
     }
 }
 
 // ----------------------------------------------------------------------------
-// Step 5 — Implement MachineApi (mandatory plumbing + your mutation logic).
+// Step 5 — MachineApi: message dispatch + mutation handling.
+// The four arms below are required boilerplate. Only `HttpApiJsonRequest` /
+// `api_mutate` need machine-specific code.
 // ----------------------------------------------------------------------------
 impl MachineApi for MyMachine {
-    fn api_get_sender(&self) -> smol::channel::Sender<MachineMessage> {
-        self.api_sender.clone()
+    fn act_machine_message(&mut self, msg: MachineMessage) {
+        match msg {
+            MachineMessage::SubscribeNamespace(namespace) => {
+                self.namespace.namespace = Some(namespace);
+                // Push current state so the new subscriber doesn't wait for
+                // the next ~33 ms emit cycle.
+                self.emit_state();
+            }
+            MachineMessage::UnsubscribeNamespace => {
+                self.namespace.namespace = None;
+            }
+            MachineMessage::HttpApiJsonRequest(value) => {
+                let _res = self.api_mutate(value);
+            }
+            MachineMessage::RequestValues(sender) => {
+                sender
+                    .send(MachineValues {
+                        state: serde_json::to_value(self.get_state())
+                            .expect("Failed to serialize state"),
+                        live_values: serde_json::Value::Null,
+                    })
+                    .expect("Failed to send values");
+            }
+        }
     }
 
-    fn api_mutate(&mut self, request_body: Value) -> Result<(), anyhow::Error> {
-        // TODO: deserialize and dispatch mutations:
-        //
-        // let mutation: Mutation = serde_json::from_value(request_body)?;
-        // match mutation {
-        //     Mutation::SetOutput { index, on } => self.set_output(index, on),
-        //     Mutation::SetAllOutputs { on } => { /* ... */ }
-        // }
+    fn get_api_sender(&self) -> tokio::sync::mpsc::Sender<MachineMessage> {
+        self.sender.clone()
+    }
 
-        // If this machine is read-only, just return Ok(()) without parsing:
-        let _ = request_body;
+    fn api_mutate(&mut self, value: Value) -> Result<(), anyhow::Error> {
+        // For read-only machines, delete the lines below and just `Ok(())`.
+        let mutation: Mutation = serde_json::from_value(value)?;
+        match mutation {
+            // TODO: dispatch arms, e.g.:
+            // Mutation::SetOutput { index, value } => self.set_output(index, value),
+            #[allow(unreachable_patterns)]
+            _ => {}
+        }
         Ok(())
     }
