+++ machine_implementations/src/minimal_machines/BOILERPLATE_TEMPLATE/mod.rs
@@ -1,23 +1,25 @@
 // ============================================================================
-// mod.rs — Core struct definition
+// mod.rs — Core struct definition + Machine trait + business logic
 // ============================================================================
-// This file defines the machine struct and implements the `Machine` trait.
-//
 // FIND & REPLACE to adapt this template:
-//   MyMachine      → YourMachineName  (e.g. WagoDiMachine)
-//   MY_MACHINE_ID  → your constant    (e.g. WAGO_DI_MACHINE)
+//   MyMachine      → YourMachineName  (e.g. Wago750_460Machine)
+//   MY_MACHINE_ID  → your constant    (e.g. WAGO_750_460_MACHINE)
+//   my_machine     → your_machine     (used for namespace types — match dir name)
+//
+// Files in this directory:
+//   mod.rs — this file: struct, MACHINE_IDENTIFICATION, helper methods
+//   new.rs — MachineNew trait: hardware initialization, called once at startup
+//   act.rs — Machine trait: act(), react(), get_identification()
+//   api.rs — MachineApi trait + StateEvent + Mutation + Namespace
 // ============================================================================
 
 use std::time::Instant;
 
-use control_core::socketio::namespace::NamespaceCacheingLogic;
-use smol::channel::{Receiver, Sender};
-
-use crate::{
-    AsyncThreadMessage, Machine, MachineMessage, MY_MACHINE_ID, VENDOR_QITECH,
-    machine_identification::{MachineIdentification, MachineIdentificationUnique},
-};
 use self::api::{MyMachineEvents, MyMachineNamespace, StateEvent};
+use crate::{MY_MACHINE_ID, MachineMessage, QiTechMachine, VENDOR_QITECH};
+use control_core::socketio::namespace::NamespaceCacheingLogic;
+use qitech_lib::machines::{MachineIdentification, MachineIdentificationUnique};
+use tokio::sync::mpsc::{Receiver, Sender};
 
 pub mod act;
 pub mod api;
@@ -26,78 +28,66 @@ pub mod new;
 // ----------------------------------------------------------------------------
 // Step 1 — Define your machine struct.
 //
-// Add the hardware handles your machine needs (DigitalOutput, DigitalInput,
-// AnalogInput, etc.) as fields. The first five fields are mandatory plumbing
-// for every machine — do not remove them.
+// The first 5 fields are mandatory plumbing — every minimal machine has them.
+// Add your hardware handles + any per-machine state below.
+//
+// Hardware handle conventions:
+//   - Beckhoff terminal:        Rc<RefCell<EL2004>>           or
+//                               Rc<RefCell<dyn DigitalOutputDevice>>  (trait obj)
+//   - WAGO module (subdevice):  Box<Wago750_530>              (owned by machine)
 // ----------------------------------------------------------------------------
 #[derive(Debug)]
 pub struct MyMachine {
-    // --- mandatory plumbing (keep as-is) ------------------------------------
-    pub api_receiver: Receiver<MachineMessage>,
-    pub api_sender: Sender<MachineMessage>,
+    // --- mandatory plumbing -------------------------------------------------
+    pub receiver: Receiver<MachineMessage>,
+    pub sender: Sender<MachineMessage>,
     pub machine_identification_unique: MachineIdentificationUnique,
-    pub main_sender: Option<Sender<AsyncThreadMessage>>,
     pub namespace: MyMachineNamespace,
     pub last_state_emit: Instant,
-
-    // --- TODO: add your hardware handles here --------------------------------
-    // Examples:
-    //   pub douts: [DigitalOutput; 4],
-    //   pub dins:  [DigitalInput; 4],
-    //   pub ain:   AnalogInput,
+    // --- TODO: add hardware handles + state fields here ---------------------
+    // pub digital_output_device: Rc<RefCell<dyn DigitalOutputDevice>>,
+    // pub last_output_state: [bool; 4],
 }
 
-// ----------------------------------------------------------------------------
-// Step 2 — Implement the `Machine` trait (mandatory, do not change).
-// ----------------------------------------------------------------------------
-impl Machine for MyMachine {
-    fn get_machine_identification_unique(&self) -> MachineIdentificationUnique {
-        self.machine_identification_unique.clone()
-    }
-
-    fn get_main_sender(&self) -> Option<Sender<AsyncThreadMessage>> {
-        self.main_sender.clone()
-    }
-}
+impl QiTechMachine for MyMachine {}
 
 // ----------------------------------------------------------------------------
-// Step 3 — Set the unique machine ID constant.
+// Step 2 — Bind MACHINE_IDENTIFICATION.
 //
-// The constant MY_MACHINE_ID must be declared in machines/src/lib.rs and must
-// be a unique u16 value not used by any other machine. See lib.rs for the
-// existing values and pick the next available one.
+// MY_MACHINE_ID must be declared in machine_implementations/src/lib.rs and
+// must be a unique u16. See lib.rs and pick the next free hex value.
 // ----------------------------------------------------------------------------
 impl MyMachine {
     pub const MACHINE_IDENTIFICATION: MachineIdentification = MachineIdentification {
         vendor: VENDOR_QITECH,
         machine: MY_MACHINE_ID,
     };
-}
 
-// ----------------------------------------------------------------------------
-// Step 4 — Add your machine's business logic methods here.
-//
-// `get_state()` and `emit_state()` are called by act.rs and api.rs — keep
-// their signatures stable.  Add any other helpers your machine needs.
-// ----------------------------------------------------------------------------
-impl MyMachine {
-    /// Build the current state snapshot (used for serialization and events).
+    // ------------------------------------------------------------------------
+    // Step 3 — Build the state snapshot used for emit + HTTP polling.
+    // Called from act.rs (periodic) and api.rs (RequestValues).
+    // ------------------------------------------------------------------------
     pub fn get_state(&self) -> StateEvent {
         StateEvent {
-            // TODO: fill with your actual state fields
+            // TODO: fill from your hardware + struct state
         }
     }
 
-    /// Serialize and broadcast the current state to all subscribers.
     pub fn emit_state(&mut self) {
         let event = self.get_state().build();
         self.namespace.emit(MyMachineEvents::State(event));
     }
 
-    // TODO: add domain-specific helpers, e.g.:
+    // ------------------------------------------------------------------------
+    // Step 4 — Add domain-specific helpers here.
     //
-    // pub fn set_output(&mut self, index: usize, on: bool) {
-    //     self.douts[index].set(on);
+    // Helpers usually mutate a hardware handle then call `self.emit_state()`
+    // so subscribers see the change immediately.
+    //
+    // pub fn set_output(&mut self, index: usize, value: bool) {
+    //     self.digital_output_device.borrow_mut().set_output(index, value);
+    //     self.last_output_state[index] = value;
     //     self.emit_state();
     // }
+    // ------------------------------------------------------------------------
 }
